+++ RELEASES.md
@@ -103,6 +103,7 @@ Upcoming
 - Disable monitoring by default. It's usage is mostly replaced by log management tools like Splunk and Logstash. When monitoring is enabled keep in mind that it will cause OutOfMemory exceptions after a number of configuration reloads
 - Prevent OutOfMemory exception after a number of configuration reloads
 - Add parameter types "map" and "list" to Larva TestTool
+- Add possibility to use zip files with multiple jar files in Upload Configuration